pyproject.toml
[tool.poetry]
name = "powerapps-time-entry-cli"
version = "0.1.0"
description = "Interactive CLI to generate, validate, preview, and submit monthly Power Apps time entries"
authors = ["Nurzhanat Zhussup <you@example.com>"]
readme = "README.md"
packages = [{ include = "powerapps_time_cli", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.3"
rich = "^13.7.1"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.5.0"
mypy = "^1.10.0"

[tool.poetry.scripts]
powerapps-time = "powerapps_time_cli.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
